// studentadminportal-ui/src/app/models/api-models/student.model.ts
import { Address } from './address.model';
import { Gender } from './gender.model';

export interface Student {
  id: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  email: string;
  mobile: number;
  profileImageUrl: string;
  genderId: string;
  gender: Gender;
  address: Address;
}
export interface StudentList {
  data: Student[];
}
export interface StudentDetail {
  data: Student;
}